Converter docs (#55)

+++ src/EfficientDynamoDb/Internal/Converters/Primitives/Numbers/IntDdbConverter.cs
@@ -4,6 +4,7 @@ using System.Text.Json;
 using EfficientDynamoDb.Context;
 using EfficientDynamoDb.DocumentModel;
 using EfficientDynamoDb.DocumentModel.AttributeValues;
+using EfficientDynamoDb.DocumentModel.Constants;
 using EfficientDynamoDb.DocumentModel.Converters;
 using EfficientDynamoDb.DocumentModel.Exceptions;
 using EfficientDynamoDb.DocumentModel.Extensions;
@@ -15,7 +16,7 @@ namespace EfficientDynamoDb.Internal.Converters.Primitives.Numbers
 {
     internal sealed class IntDdbConverter : NumberDdbConverter<int>, IDictionaryKeyConverter<int>, ISetValueConverter<int>
     {
-        public override int Read(in AttributeValue attributeValue) => attributeValue.AsNumberAttribute().ToByte();
+        public override int Read(in AttributeValue attributeValue) => attributeValue.AsNumberAttribute().ToInt();
 
         public override void Write(in DdbWriter writer, ref int value) => WriteInlined(writer.JsonWriter, ref value);
